+++ ui-game/src/main/java/com/jayasuryat/uigame/composable/toggle/Toggle.kt
@@ -35,7 +35,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.jayasuryat.minesweeperui.R
-import com.jayasuryat.uigame.logic.ToggleState
+import com.jayasuryat.uigame.data.model.ToggleState
 import com.jayasuryat.util.LogCompositions
 
 @OptIn(ExperimentalAnimationApi::class)
