+++ ui-game/src/main/java/com/jayasuryat/uigame/composable/topbar/started/GameProgressChip.kt
@@ -36,7 +36,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.IntOffset
 import androidx.compose.ui.unit.dp
-import com.jayasuryat.uigame.logic.GameProgress
+import com.jayasuryat.uigame.logic.model.GameProgress
 import com.jayasuryat.util.LogCompositions
 
 @OptIn(ExperimentalAnimationApi::class)
